+++ app/helpers/related_urls_helper.rb
@@ -100,152 +100,6 @@ module RelatedUrlsHelper
     }
   }
 
-  related_url_content_type_map_for_umm_s = {
-    'CollectionURL' => {
-      'text' => 'Collection URL',
-      'types' => {
-        'DATA SET LANDING PAGE' => {
-          'text' => 'Data Set Landing Page',
-          'subtypes' => []
-        },
-        'DOI' => {
-          'text' => 'DOI',
-          'subtypes' => []
-        },
-        'EXTENDED METADATA' => {
-          'text' => 'Extended Metadata',
-          'subtypes' => []
-        },
-        'PROFESSIONAL HOME PAGE' => {
-          'text' => 'Professional Home Page',
-          'subtypes' => []
-        },
-        'PROJECT HOME PAGE' => {
-          'text' => 'Project Home Page',
-          'subtypes' => []
-        }
-      }
-    },
-    'PublicationURL' => {
-      'text' => 'Publication URL',
-      'types' => {
-        'VIEW RELATED INFORMATION' => {
-          'text' => 'View Related Information',
-          'subtypes' => [
-            ['Algorithm Theoretical Basis Document', 'ALGORITHM THEORETICAL BASIS DOCUMENT'],
-            ['Calibration Data Documentation', 'CALIBRATION DATA DOCUMENTATION'],
-            ['Case Study', 'CASE STUDY'],
-            ['Data Quality', 'DATA QUALITY'],
-            ['Data Usage', 'DATA USAGE'],
-            ['Deliverables Checklist', 'DELIVERABLES CHECKLIST'],
-            ['General Documentation', 'GENERAL DOCUMENTATION'],
-            ['How To', 'HOW-TO'],
-            ['PI Documentation', 'PI DOCUMENTATION'],
-            ['Processing History', 'PROCESSING HISTORY'],
-            ['Production Version History', 'PRODUCTION VERSION HISTORY'],
-            ['Product Quality Assessment', 'PRODUCT QUALITY ASSESSMENT'],
-            ['Product Usage', 'PRODUCT USAGE'],
-            ['Product History', 'PRODUCT HISTORY'],
-            ['Publications', 'PUBLICATIONS'],
-            ['Radiometric And Geometric Calibration Methods', 'RADIOMETRIC AND GEOMETRIC CALIBRATION METHODS'],
-            ['Read Me', 'READ-ME'],
-            ['Recipe', 'RECIPE'],
-            ['Requirements And Design', 'REQUIREMENTS AND DESIGN'],
-            ['Science Data Product Software Documentation', 'SCIENCE DATA PRODUCT SOFTWARE DOCUMENTATION'],
-            ['Science Data Product Validation', 'SCIENCE DATA PRODUCT VALIDATION'],
-            ['User Feedback', 'USER FEEDBACK'],
-            ["User's Guide", "USER'S GUIDE"]
-          ]
-        }
-      }
-    },
-    'DataCenterURL' => {
-      'text' => 'Data Center URL',
-      'types' => {
-        'HOME PAGE' => {
-          'text' => 'Home Page',
-          'subtypes' => []
-        }
-      }
-    },
-    'DataContactURL' => {
-      'text' => 'Data Contact URL',
-      'types' => {
-        'HOME PAGE' => {
-          'text' => 'Home Page',
-          'subtypes' => []
-        }
-      }
-    },
-    'DistributionURL' => {
-      'text' => 'Distribution URL',
-      'types' => {
-        'GET DATA' => {
-          'text' => 'Get Data',
-          'subtypes' => [
-            ['Datacast URL', 'DATACAST URL'],
-            ['Earthdata Search', 'EARTHDATA SEARCH'],
-            ['ECHO', 'ECHO'],
-            ['EDG', 'EDG'],
-            ['EOSDIS Data Pool', 'EOSDIS DATA POOL'],
-            ['GDS', 'GDS'],
-            ['Giovanni', 'GIOVANNI'],
-            ['KML', 'KML'],
-            ['LAADS', 'LAADS'],
-            ['LANCE', 'LANCE'],
-            ['LAS', 'LAS'],
-            ['Mirador', 'MIRADOR'],
-            ['MODAPS', 'MODAPS'],
-            ['NOAA Class', 'NOAA CLASS'],
-            ['On Line Archive', 'ON-LINE ARCHIVE'],
-            ['Reverb', 'REVERB']
-          ]
-        },
-        'GET SERVICE' => {
-          'text' => 'Get Service',
-          'subtypes' => [
-            ['Access Map Viewer', 'ACCESS MAP VIEWER'],
-            ['Access Mobile App', 'ACCESS MOBILE APP'],
-            ['Access Web Service', 'ACCESS WEB SERVICE'],
-            ['DIF', 'DIF'],
-            ['Map Service', 'MAP SERVICE'],
-            ['NOMADS', 'NOMADS'],
-            ['OPeNDAP Data', 'OPENDAP DATA'],
-            ['OPeNDAP Data (Dods)', 'OPENDAP DATA (DODS)'],
-            ['OPeNDAP Directory (Dods)', 'OPENDAP DIRECTORY (DODS)'],
-            ['Open Search', 'OpenSearch'],
-            ['SERF', 'SERF'],
-            ['Software Package', 'SOFTWARE PACKAGE'],
-            ['SSW', 'SSW'],
-            ['Subsetter', 'SUBSETTER'],
-            ['Thredds Catalog', 'THREDDS CATALOG'],
-            ['Thredds Data', 'THREDDS DATA'],
-            ['Thredds Directory', 'THREDDS DIRECTORY'],
-            ['Tool','TOOL'],
-            ['Web Coverage Service (WCS)', 'WEB COVERAGE SERVICE (WCS)'],
-            ['Web Feature Service (WFS)', 'WEB FEATURE SERVICE (WFS)'],
-            ['Web Map For Time Series', 'WEB MAP FOR TIME SERIES'],
-            ['Web Map Service (WMS)', 'WEB MAP SERVICE (WMS)'],
-            ['Workflow (Service Chain)', 'WORKFLOW (SERVICE CHAIN)']
-          ]
-        }
-      }
-    },
-    'VisualizationURL' => {
-      'text' => 'Visualization URL',
-      'types' => {
-        'GET RELATED VISUALIZATION' => {
-          'text' => 'Get Related Visualization',
-          'subtypes' => [
-            ['GIBS', 'GIBS'],
-            ['Giovanni', 'GIOVANNI'],
-            ['Map', 'MAP']
-          ]
-        }
-      }
-    }
-  }
-
   related_url_content_type_map = {
     'CollectionURL' => {
       'text' => 'Collection URL',
@@ -403,7 +257,6 @@ module RelatedUrlsHelper
   # TODO: need to verify if all these options are necessary if they get changed
   # by our coffee files through the mappings.
   UMMCRelatedURLContentTypeMap = related_url_content_type_map.deep_dup
-  UMMSRelatedURLContentTypeMap = related_url_content_type_map_for_umm_s
   UMMTURLContentTypeMap = url_content_type_map_for_umm_t
   UMMTRelatedURLContentTypeMap = related_url_content_type_map_for_umm_t
 
@@ -412,11 +265,6 @@ module RelatedUrlsHelper
     ['Data Center URL', 'DataCenterURL'],
     ['Data Contact URL', 'DataContactURL'],
     ['Distribution URL', 'DistributionURL'],
-    ['Publication URL', 'PublicationURL'],
-    ['Visualization URL', 'VisualizationURL']
-  ]
-  UMMSRelatedURLContentTypeOptions = [
-    ['Collection URL', 'CollectionURL'],
     ['Publication URL', 'PublicationURL'],
     ['Visualization URL', 'VisualizationURL']
   ]
@@ -452,18 +300,6 @@ module RelatedUrlsHelper
     ['Professional Home Page', 'PROFESSIONAL HOME PAGE'],
     ['Project Home Page', 'PROJECT HOME PAGE'],
     ['Use Service API', 'USE SERVICE API'],
-    ['View Related Information', 'VIEW RELATED INFORMATION']
-  ]
-  UMMSRelatedURLTypeOptions = [
-    ['Get Data', 'GET DATA'],
-    ['Data Set Landing Page', 'DATA SET LANDING PAGE'],
-    ['DOI', 'DOI'],
-    ['Extended Metadata', 'EXTENDED METADATA'],
-    ['Get Related Visualization', 'GET RELATED VISUALIZATION'],
-    ['Get Service', 'GET SERVICE'],
-    ['Home Page', 'HOME PAGE'],
-    ['Professional Home Page', 'PROFESSIONAL HOME PAGE'],
-    ['Project Home Page', 'PROJECT HOME PAGE'],
     ['View Related Information', 'VIEW RELATED INFORMATION']
   ]
   UMMTRelatedURLTypeOptions = [
@@ -549,71 +385,6 @@ module RelatedUrlsHelper
     ['Web Map Service (WMS)', 'WEB MAP SERVICE (WMS)'],
     ['Web Map Tile Service (WMTS)', 'WEB MAP TILE SERVICE (WMTS)'],
     ['Worldview', 'WORLDVIEW'],
-    ["User's Guide", "USER'S GUIDE"]
-  ]
-  UMMSRelatedURLSubtypeOptions = [
-    ['Access Map Viewer', 'ACCESS MAP VIEWER'],
-    ['Access Mobile App', 'ACCESS MOBILE APP'],
-    ['Access Web Service', 'ACCESS WEB SERVICE'],
-    ['Algorithm Theoretical Basis Document', 'ALGORITHM THEORETICAL BASIS DOCUMENT'],
-    ['Calibration Data Documentation', 'CALIBRATION DATA DOCUMENTATION'],
-    ['Case Study', 'CASE STUDY'],
-    ['Data Quality', 'DATA QUALITY'],
-    ['Data Usage', 'DATA USAGE'],
-    ['Datacast URL', 'DATACAST URL'],
-    ['Deliverables Checklist', 'DELIVERABLES CHECKLIST'],
-    ['DIF', 'DIF'],
-    ['Earthdata Search', 'EARTHDATA SEARCH'],
-    ['ECHO', 'ECHO'],
-    ['EDG', 'EDG'],
-    ['EOSDIS Data Pool', 'EOSDIS DATA POOL'],
-    ['GDS', 'GDS'],
-    ['General Documentation', 'GENERAL DOCUMENTATION'],
-    ['GIBS', 'GIBS'],
-    ['Giovanni', 'GIOVANNI'],
-    ['How To', 'HOW-TO'],
-    ['KML', 'KML'],
-    ['LAADS', 'LAADS'],
-    ['LANCE', 'LANCE'],
-    ['LAS', 'LAS'],
-    ['Map Service', 'MAP SERVICE'],
-    ['Map', 'MAP'],
-    ['Mirador', 'MIRADOR'],
-    ['MODAPS', 'MODAPS'],
-    ['NOAA Class', 'NOAA CLASS'],
-    ['NOMADS', 'NOMADS'],
-    ['On Line Archive', 'ON-LINE ARCHIVE'],
-    ['Open Search', 'OpenSearch'],
-    ['OPeNDAP Data (Dods)', 'OPENDAP DATA (DODS)'],
-    ['OPeNDAP Data', 'OPENDAP DATA'],
-    ['OPeNDAP Directory (Dods)', 'OPENDAP DIRECTORY (DODS)'],
-    ['PI Documentation', 'PI DOCUMENTATION'],
-    ['Processing History', 'PROCESSING HISTORY'],
-    ['Product History', 'PRODUCT HISTORY'],
-    ['Product Quality Assessment', 'PRODUCT QUALITY ASSESSMENT'],
-    ['Product Usage', 'PRODUCT USAGE'],
-    ['Production Version History', 'PRODUCTION VERSION HISTORY'],
-    ['Publications', 'PUBLICATIONS'],
-    ['Radiometric And Geometric Calibration Methods', 'RADIOMETRIC AND GEOMETRIC CALIBRATION METHODS'],
-    ['Read Me', 'READ-ME'],
-    ['Recipe', 'RECIPE'],
-    ['Requirements And Design', 'REQUIREMENTS AND DESIGN'],
-    ['Reverb', 'REVERB'],
-    ['Science Data Product Software Documentation', 'SCIENCE DATA PRODUCT SOFTWARE DOCUMENTATION'],
-    ['Science Data Product Validation', 'SCIENCE DATA PRODUCT VALIDATION'],
-    ['SERF', 'SERF'],
-    ['Software Package', 'SOFTWARE PACKAGE'],
-    ['SSW', 'SSW'],
-    ['Subsetter', 'SUBSETTER'],
-    ['Threads Catalog', 'THREADS CATALOG'],
-    ['Threads Data', 'THREADS DATA'],
-    ['Thredds Directory', 'THREDDS DIRECTORY'],
-    ['User Feedback', 'USER FEEDBACK'],
-    ['Web Coverage Service (WCS)', 'WEB COVERAGE SERVICE (WCS)'],
-    ['Web Feature Service (WFS)', 'WEB FEATURE SERVICE (WFS)'],
-    ['Web Map For Time Series', 'WEB MAP FOR TIME SERIES'],
-    ['Web Map Service (WMS)', 'WEB MAP SERVICE (WMS)'],
-    ['Workflow (Service Chain)', 'WORKFLOW (SERVICE CHAIN)'],
     ["User's Guide", "USER'S GUIDE"]
   ]
   UMMTRelatedURLSubtypeOptions = [
